# Drop extra string held in relabel.Regexp struct