--- a/model/relabel/relabel.go
+++ b/model/relabel/relabel.go
@@ -142,17 +142,13 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 // Regexp encapsulates a regexp.Regexp and makes it YAML marshalable.
 type Regexp struct {
 	*regexp.Regexp
-	original string
 }
 
 // NewRegexp creates a new anchored Regexp and returns an error if the
 // passed-in regular expression does not compile.
 func NewRegexp(s string) (Regexp, error) {
 	regex, err := regexp.Compile("^(?:" + s + ")$")
-	return Regexp{
-		Regexp:   regex,
-		original: s,
-	}, err
+	return Regexp{Regexp: regex}, err
 }
 
 // MustNewRegexp works like NewRegexp, but panics if the regular expression does not compile.
@@ -180,8 +176,11 @@ func (re *Regexp) UnmarshalYAML(unmarshal func(interface{}) error) error {
 
 // MarshalYAML implements the yaml.Marshaler interface.
 func (re Regexp) MarshalYAML() (interface{}, error) {
-	if re.original != "" {
-		return re.original, nil
+	if re.String() != "" {
+		str := re.String()
+		str = strings.TrimPrefix(str, "^(?:")
+		str = strings.TrimSuffix(str, ")$")
+		return str, nil
 	}
 	return nil, nil
 }