--- a/model/relabel/relabel.go
+++ b/model/relabel/relabel.go
@@ -177,12 +177,16 @@ func (re *Regexp) UnmarshalYAML(unmarshal func(interface{}) error) error {
 // MarshalYAML implements the yaml.Marshaler interface.
 func (re Regexp) MarshalYAML() (interface{}, error) {
 	if re.String() != "" {
-		str := re.String()
-		str = strings.TrimPrefix(str, "^(?:")
-		str = strings.TrimSuffix(str, ")$")
-		return str, nil
+		return re.String(), nil
 	}
 	return nil, nil
+}
+
+// String returns the original string used to compile the regular expression.
+func (re Regexp) String() string {
+	str := re.Regexp.String()
+	// Trim the anchor `^(?:` prefix and `)$` suffix.
+	return str[4 : len(str)-2]
 }
 
 // Process returns a relabeled copy of the given label set. The relabel configurations