Signed-off-by: Paschalis Tsilias <paschalist0@gmail.com>

This was initially pointed out by @rfratto when experimenting with something on the Agent. Looking at [grafana/regexp](https://github.com/grafana/regexp/blob/2e8d9baf4ac26b43e1eb1e8a62fb920a6efaef81/regexp.go#L82-L107), the result of the regexp.String() call returns what was passed to Compile when creating the Regexp object. 

Trimming the prefix and suffix manually would give us the same string back, trading some extra instructions for memory as we won't have to keep around an extra copy of the original string around anymore.

Do you feel that's worth it as a change?


## Benchmarks
I've run [this](https://github.com/tpaschalis/prometheus/commit/34a6ce345223c42b3735d7ea2406587c22948e76) benchmark on main (f2ba2a0), and then on main with this change applied.

While the results aren't spectacular, they showed some slight improvement even for a single relabel_config step.

```
$ go test ./... -run=None -bench=. -benchmem -benchtime=5s -count=10 > before
$ go test ./... -run=None -bench=. -benchmem -benchtime=5s -count=10 > after

$ benchstat before after
name                  old time/op    new time/op    delta
RegexpDropOriginal-8    18.1µs ± 2%    17.6µs ± 0%  -2.90%  (p=0.000 n=9+10)

name                  old alloc/op   new alloc/op   delta
RegexpDropOriginal-8    27.6kB ± 0%    27.5kB ± 0%  -0.33%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
RegexpDropOriginal-8       230 ± 0%       227 ± 0%  -1.30%  (p=0.000 n=10+10)
```