@bboreham 2022-06-09
For historical perspective, this is a reversion of the change made here:

https://github.com/prometheus/prometheus/pull/1044/files#diff-fe44f09c4d5977b5f5eaea29170b6a0748819c9d02271746a20d81a5f3efca17R768

Maybe @juliusv remembers why this wasn't done?  
Maybe didn't want to assume we could reverse the original out from the modified regexp?