@roidelapluie 2022-06-09
> String returns the source text used to compile the regular expression.

From go documentation.