@juliusv 2022-06-13
> Maybe @juliusv remembers why this wasn't done?
Maybe didn't want to assume we could reverse the original out from the modified regexp?

Something like that, yeah. I probably preferred the more explicit storing of the original and it seemed safer. But it seems like `String()` really returns the original regex as well, so this PR should be fine as well.

:+1: